Adjust ImporterDirectoryWatcherTest.java to confirm directory creation instead of trying to run .run()
It doesn't make much sense to test the scheduledExecutor when we already have a test suite for the ImporterProcessor called by the scheduled executor. JMockit doesn't play nicely (easily) with Spring configuration objects, so in the interest of time I refactored this test suite to confirm that we create the necessary directories without the use of JMockit.