comment out outputSchemaVersion in OdeProperties and add Import annotation for BuildProperties.class to allow usage of SpringBootTest annotations.
BuildProperties wasn't found unless mvn package was run prior to testing. That's slow and not intuitive, so the annotation fixed the issue. Commenting out outputSchemaVersion allowed for setting OdeProperties as a Component instead of a Configuration class, since that's essentially what it is, and we don't need output schema to be served from a bean in the long run. It can be used as a Value annotated field wherever needed instead