test: refactored test to use unique keys for consumed matching
Since this test suite produces to and consumes from the same topics in multiple tests, we needed a better way than `getSingleRecord` to perform test assertions. By using uuids as keys, we are able to reuse the same topics for multiple tests and select the correct records for test assertions without risk of tests conflicting with one another.