Allow Kafka to be shared outside of ODE Network (#225)
The Bitnami Kafka images used by the ODE only advertise a small selection of ports to listen for new data from applications. This is set in the jpo-utils docker-compose-kafka.yml file here

Specifically: KAFKA_CFG_ADVERTISED_LISTENERS: "PLAINTEXT://kafka:9094,EXTERNAL://${KAFKA_BOOTSTRAP_SERVERS:-kafka:9092}"

In the default configuration, this causes the Kafka broker to advertise to connected applications that they should send data to Kafka:9092. For applications such as the ADM and AEM modules this works fine, as they are typically deployed alongside the ODE itself and using a container ID of Kafka will properly route back to the Bitnami container. However, this strategy fails if another container or application outside of the Docker-compose Network tries to connect. For example, spinning up a separate copy of the geojson converter will fail since it cannot route properly back to the hostname Kafka.

This PR solves this issue by directly specifying the KAFKA_BOOTSTRAP_SERVERS env var. This changes the default advertised broker location to be DOCKER_HOST_IP:9092 which is typically routable by external containers.